Struct conflux::cli::Cli

pub struct Cli {

    pub mode: Option<String>,
    pub port: Option<String>,
    pub udp_port: Option<String>,
    pub jsonrpc_ws_port: Option<String>,
    pub jsonrpc_tcp_port: Option<String>,
    pub jsonrpc_http_port: Option<String>,
    pub jsonrpc_cors: Option<String>,
    pub jsonrpc_http_keep_alive: Option<String>,
    pub log_file: Option<String>,
    pub log_level: Option<String>,
    pub log_conf: Option<String>,
    pub config: Option<String>,
    pub bootnodes: Option<String>,
    pub netconf_dir: Option<String>,
    pub public_address: Option<String>,
    pub net_key: Option<String>,
    pub start_mining: Option<String>,
    pub mining_author: Option<String>,
    pub ledger_cache_size: Option<String>,
    pub db_cache_size: Option<String>,
    pub enable_discovery: Option<String>,
    pub node_table_timeout_s: Option<String>,
    pub node_table_promotion_timeout_s: Option<String>,
    pub test_mode: Option<String>,
    pub db_compact_profile: Option<String>,
    pub block_db_dir: Option<String>,
    pub load_test_chain: Option<String>,
    pub egress_queue_capacity: Option<String>,
    pub egress_min_throttle: Option<String>,
    pub egress_max_throttle: Option<String>,
    pub get_logs_epoch_batch_size: Option<String>,
    pub get_logs_filter_max_epoch_range: Option<String>,
    pub get_logs_filter_max_limit: Option<String>,
    pub get_logs_filter_max_block_number_range: Option<String>,
    pub account_provider_refresh_time_ms: Option<String>,
    pub dev_pos_private_key_encryption_password: Option<String>,
    pub pos_started_as_voter: Option<String>,
    pub light: bool,
    pub archive: bool,
    pub tg_archive: bool,
    pub full: bool,
    pub command: Option<Commands>,

}

Conflux client

Fields
mode: Option<String>

Use the preset testing configurations. dev or test.

port: Option<String>

Specify the port for P2P connections.

udp_port: Option<String>

Specify the UDP port for peer discovery.

jsonrpc_ws_port: Option<String>

Specify the port for the WebSocket JSON-RPC API server.

jsonrpc_tcp_port: Option<String>

Specify the port for the TCP JSON-RPC API server.

jsonrpc_http_port: Option<String>

Specify the port for the HTTP JSON-RPC API server.

jsonrpc_cors: Option<String>

Specify CORS header for HTTP JSON-RPC API responses.

jsonrpc_http_keep_alive: Option<String>

Enable HTTP/1.1 keep alive header.

log_file: Option<String>

Specify the filename for the log. Stdout will be used by default if omitted.

log_level: Option<String>

Can be error/warn/info/debug/trace. Default is the info level.

log_conf: Option<String>

Sets a custom log config file.

config: Option<String>

Sets a custom config file.

bootnodes: Option<String>

Sets a custom list of bootnodes.

netconf_dir: Option<String>

Sets a custom directory for network configurations.

public_address: Option<String>

Sets a custom public address to be connected by others.

net_key: Option<String>

Sets a custom secret key to generate unique node ID.

start_mining: Option<String>

Start mining if set to true. Ensure that mining-author is set.

mining_author: Option<String>

Set the address to receive mining rewards.

ledger_cache_size: Option<String>

Sets the ledger cache size.

db_cache_size: Option<String>

Sets the db cache size.

enable_discovery: Option<String>

Enable discovery protocol.

node_table_timeout_s: Option<String>

How often Conflux updates its peer table (default 300).

node_table_promotion_timeout_s: Option<String>

How long Conflux waits for promoting a peer to trustworthy (default 3 * 24 * 3600).

test_mode: Option<String>

Sets test mode for adding latency

db_compact_profile: Option<String>

Sets the compaction profile of RocksDB.

block_db_dir: Option<String>

Sets the root path of db.

load_test_chain: Option<String>

Sets the test chain json file.

egress_queue_capacity: Option<String>

Sets egress queue capacity of P2P network.

egress_min_throttle: Option<String>

Sets minimum throttling queue size of egress.

egress_max_throttle: Option<String>

Sets maximum throttling queue size of egress.

get_logs_epoch_batch_size: Option<String>

Sets the size of the epoch batches used during log filtering.

get_logs_filter_max_epoch_range: Option<String>

Sets the maximum number of allowed epochs during log filtering.

get_logs_filter_max_limit: Option<String>

Sets the maximum number of log entries returned during log filtering.

get_logs_filter_max_block_number_range: Option<String>

Sets the maximum number of allowed blocks during log filtering.

account_provider_refresh_time_ms: Option<String>

Sets the time after which accounts are re-read from disk.

dev_pos_private_key_encryption_password: Option<String>

Sets the encryption password for the pos private key file. It’s used to encrypt a new generated key or to decrypt an existing key file.

pos_started_as_voter: Option<String>

If true, the node will start PoS election and voting if it’s available.

light: boolarchive: booltg_archive: boolfull: boolcommand: Option<Commands>